requirements.txt
Django==4.2.10
requests==2.27.1
djangorestframework==3.14.0
django-cors-headers==3.3.0
django-cors-middleware==1.3.1